Updated uint32_t i to uint64_t i in Decoder::deconstructHeader

+++ src/huffman_decoder.cpp
@@ -60,7 +60,7 @@ namespace huffman {
 
         std::string ver = "";
 
-        uint32_t i;
+        uint64_t i;
         
         // Decode version name + number (4 Bytes)
         for (i = 0; i < 4; ++i)
